improvements and unit test added
Create computer, fill the id field of the returned computer.\nDelete computer, only id requierd.\nCreate and update computer, check introduced date less than discontunued date before operatin.\nBeans id pass from int to long.\nHashcode method added.\Create and update computer with company id 0 now operating with NULL company id in DB.

+++ src/org/excilys/cli/Page.java
@@ -24,7 +24,7 @@ public class Page<T> {
 		} else {
 			this.entities = entities;
 		}
-		this.totalPage = (int) Math.ceil(entities.size() / eltByPage);
+		this.totalPage = (int) Math.ceil(entities.size() / (double) eltByPage);
 		// Handle beigining of set
 		if (totalPage == 1) {
 			this.currentEntities = entities.subList(0, currentPage * eltByPage);
